packages/desktop/pyproject.toml
[project]
name = "photo-grouper"
version = "0.1.0"
description = "this is a photo grouper"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.1",
    "networkx>=3.0",
    "numpy==2.0.0",
    "pillow>=10.0.0",
    "pillow-heif>=0.13.0",
    "pyinstaller>=6.15.0",
    "pyside6>=6.5.0",
    "onnx>=1.15.0",
    "onnxruntime>=1.16.0",
    "watchdog>=3.0.0",
]

[dependency-groups]
dev = [
    "tqdm>=4.67.1",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pre-commit>=3.5.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "onnxscript>=0.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py38"
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "B905",  # zip() without explicit strict parameter
]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    "venv",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__ files

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
(
  /(
      \.git
    | __pycache__
    | build
    | dist
    | venv
    | \.venv
  )/
)
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = [
    "build/",
    "dist/",
    "venv/",
    ".venv/",
]


[tool.uv.scripts]
build = "rm -rf build dist __pycache__ && pyinstaller photo-grouper.spec"

[tool.uv.export-to-onnx-all]
description = "Export all models to ONNX"
run = "scripts/export_to_onnx.py --all"
